requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
filelock==3.12.0
idna==3.4
Jinja2==3.1.2
llvmlite==0.40.0
MarkupSafe==2.1.2
more-itertools==9.1.0
mpmath==1.3.0
networkx==3.1
numba==0.57.0
numpy==1.24.3
openai-whisper @ git+https://github.com/openai/whisper.git@248b6cb124225dd263bb9bd32d060b6517e067f8
regex==2023.5.5
requests==2.30.0
sympy==1.12
tiktoken==0.3.3
torch==2.0.1
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==2.0.2
